Initial implementation of generic ThTr (for any .NET app)
