+++ ThrottlingTroll.AzureFunctions/ThrottlingTrollExtensions.cs
@@ -117,7 +117,7 @@ namespace ThrottlingTroll
         /// </summary>
         public static List<ThrottlingTrollConfig> GetThrottlingTrollConfig(this FunctionContext context)
         {
-            return (List<ThrottlingTrollConfig>)context.Items[ThrottlingTroll.ThrottlingTrollConfigsContextKey];
+            return (List<ThrottlingTrollConfig>)context.Items[ThrottlingTrollCore.ThrottlingTrollConfigsContextKey];
         }
 
         /// <summary>
@@ -125,7 +125,7 @@ namespace ThrottlingTroll
         /// </summary>
         public static List<LimitCheckResult> GetThrottlingTrollLimitCheckResults(this FunctionContext context)
         {
-            return (List<LimitCheckResult>)context.Items[ThrottlingTroll.LimitCheckResultsContextKey];
+            return (List<LimitCheckResult>)context.Items[ThrottlingTrollCore.LimitCheckResultsContextKey];
         }
 
         private static ThrottlingTrollMiddleware CreateMiddleware(FunctionContext context, ThrottlingTrollOptions opt)
@@ -135,7 +135,7 @@ namespace ThrottlingTroll
 
             if (opt.Log == null)
             {
-                var logger = context.InstanceServices.GetService<ILogger<ThrottlingTroll>>();
+                var logger = context.InstanceServices.GetService<ILogger<ThrottlingTrollCore>>();
                 opt.Log = logger == null ? null : (l, s) => logger.Log(l, s);
             }
